+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 
 ## master
 
+* Add `--cointracking` flag to format data for use with CoinTracking's Bulk
+  Imports feature.
 * Fix lower-bounds for `req`.